move file slice out of mod
